5_Applicativo/EmoSupporter/requirements.txt
Pillow==9.2.0
pystray==0.19.4
six==1.16.0

PySimpleGUI~=4.60.3
opencv-python~=4.6.0.66
numpy~=1.23.3
pip~=22.2.2
wheel~=0.37.1
keras~=2.10.0
h5py~=3.7.0
wrapt~=1.14.1
requests~=2.28.1
setuptools~=60.2.0
pytz~=2022.4
Jinja2~=3.1.2
pandas~=1.5.0
tensorflow~=2.10.0
tensorboard~=2.10.1
deepface~=0.0.75